Updated start scripts
Now divided into two different scripts, one that runs and updates, and another that just runs the tool. The run_and_update one should probably be used most of the time, but the other one is nice to have as a backup in case something breaks.

The run_and_update script now only clones the repository and updates dependencies when the local version is out of date. Additionally, the Python Venv is now being created in the modpack main folder so that the cached data doesn't get reset as often and therefore can start the program faster.